+++ CullFactory/Behaviours/DynamicCuller.cs
@@ -1,6 +1,6 @@
-﻿using System.Collections.Concurrent;
+﻿using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Threading.Tasks;
 using CullFactory.Extenders;
 using DunGen;
 using UnityEngine;
@@ -12,18 +12,14 @@ namespace CullFactory.Behaviours;
 /// </summary>
 public sealed class DynamicCuller : MonoBehaviour
 {
-    private static float CullDistance    => Plugin.Configuration.CullDistance.Value;
-    private static float SqrCullDistance => CullDistance * CullDistance;
+    private const float VanillaClipDistance = 400;
 
-    private static readonly List<ManualCameraRenderer>                 Monitors              = new();
     private static readonly ConcurrentDictionary<Tile, TileVisibility> VisibleTilesThisFrame = new();
-    private static readonly Queue<TileDepthTester>                     DepthTesterQueue      = new();
+    private static readonly List<ManualCameraRenderer>                 Monitors              = new();
+    private static readonly List<Vector3>                              CullOrigins           = new();
 
     private static List<ManualCameraRenderer> _enabledMonitors = new();
-
-    private static Vector3 _playerPosition;
-    private static float   _lastUpdateTime;
-    private static bool    _depthTested;
+    private static float                      _lastUpdateTime;
 
     private void OnEnable()
     {
@@ -37,15 +33,29 @@ public sealed class DynamicCuller : MonoBehaviour
                 continue;
 
             Monitors.Add(cameraRenderer);
+
+            Plugin.Log($"Found monitor camera \"{cameraRenderer.name}\"");
+        }
+
+        if (Math.Abs(Plugin.Configuration.CullDistance.Value - VanillaClipDistance) <= float.Epsilon * 2)
+            return;
+
+        foreach (var camera in FindObjectsByType<Camera>(FindObjectsSortMode.None))
+        {
+            if (Monitors.FindIndex(monitor => monitor.mapCamera == camera) != -1 ||
+                camera.farClipPlane                                        < Plugin.Configuration.CullDistance.Value)
+                continue;
+
+            camera.farClipPlane = Mathf.Min(camera.farClipPlane, Plugin.Configuration.CullDistance.Value);
+            Plugin.Log($"Set culling distance of \"{camera.name}\" to {Plugin.Configuration.CullDistance.Value}");
         }
     }
 
     public void Update()
     {
-        if (StartOfRound.Instance.allPlayersDead)
-            return;
-
-        if (Time.time - _lastUpdateTime < 1 / Plugin.Configuration.UpdateFrequency.Value)
+        if (!Plugin.Configuration.UseAdjacentRoomTesting.Value ||
+            StartOfRound.Instance.allPlayersDead              ||
+            Time.time - _lastUpdateTime < 1 / Plugin.Configuration.UpdateFrequency.Value)
             return;
 
         _lastUpdateTime = Time.time;
@@ -54,53 +64,61 @@ public sealed class DynamicCuller : MonoBehaviour
                               ? GameNetworkManager.Instance.localPlayerController.spectatedPlayerScript
                               : GameNetworkManager.Instance.localPlayerController;
 
-        _playerPosition  = localPlayer.transform.position;
         _enabledMonitors = Monitors.FindAll(monitor => monitor.mapCamera.enabled);
 
-        var isAnyPlayerInsideFactory = localPlayer.isInsideFactory ||
-                                       _enabledMonitors.FindIndex(monitor => monitor.targetedPlayer.isInsideFactory) != -1;
+        CullOrigins.Clear();
+        CullOrigins.Add(localPlayer.transform.position);
 
-        if (isAnyPlayerInsideFactory)
-            EstablishVisibleTiles();
+        foreach (var monitor in _enabledMonitors)
+        {
+            if (!monitor.targetedPlayer.isInsideFactory)
+                return;
+
+            CullOrigins.Add(monitor.targetedPlayer.transform.position);
+        }
+
+        if (CullOrigins.Count > 0)
+            IncludeVisibleTiles();
 
         foreach (var container in LevelGenerationExtender.MeshContainers.Values)
             container.SetVisible(VisibleTilesThisFrame.ContainsKey(container.parentTile));
     }
 
-    private static void EstablishVisibleTiles()
+    private static void IncludeVisibleTiles()
     {
         VisibleTilesThisFrame.Clear();
 
-        if (Plugin.Configuration.UseAdjacentRoomTesting.Value)
+        foreach (var tile in LevelGenerationExtender.MeshContainers.Keys)
         {
-            foreach (var tile in LevelGenerationExtender.MeshContainers.Keys)
-            {
-                if (!tile.Bounds.Contains(_playerPosition))
-                    continue;
+            var anyOriginCaptured = CullOrigins.RemoveAll(pos => tile.Bounds.Contains(pos)) > 0;
 
-                OccludeByTileBranching(tile);
-                break;
-            }
+            if (!anyOriginCaptured)
+                continue;
+
+            IncludeNearbyTiles(tile);
+            break;
         }
-
-        _depthTested = VisibleTilesThisFrame.Count != 0;
-
-        if (Plugin.Configuration.UseMultithreading.Value)
-            OccludeByDistanceParallel();
-        else
-            OccludeByDistance();
     }
 
-    private static void OccludeByTileBranching(Tile origin)
+    private static void IncludeNearbyTiles(Tile origin)
     {
-        DepthTesterQueue.Clear();
-        DepthTesterQueue.Enqueue(new TileDepthTester(origin, 0));
+        var depthTesterQueue = new Queue<TileDepthTester>();
+        var traversedTiles   = new HashSet<Tile>();
 
-        while (DepthTesterQueue.Count > 0)
+        depthTesterQueue.Clear();
+        depthTesterQueue.Enqueue(new TileDepthTester(origin, 0));
+
+        while (depthTesterQueue.Count > 0)
         {
-            var tile = DepthTesterQueue.Dequeue();
+            var tile = depthTesterQueue.Dequeue();
 
+            // We use a second list here so other depth searches won't interfere
+            // E.g. A is the first process, B is the second:
+            //       [B3]  <<- Not traversed because B sees (A2) as done and will halt early
+            // [A1]  [A2]  [A3]
+            //       [B1]
             VisibleTilesThisFrame.TryAdd(tile.tile, LevelGenerationExtender.MeshContainers[tile.tile]);
+            traversedTiles.Add(tile.tile);
 
             if (tile.iteration == Plugin.Configuration.MaxBranchingDepth.Value - 1)
                 continue;
@@ -109,48 +127,14 @@ public sealed class DynamicCuller : MonoBehaviour
             {
                 var neighborTile = doorway.ConnectedDoorway.Tile;
 
-                // This is possible due to traversing to a tile that has null doorways
-                if (VisibleTilesThisFrame.ContainsKey(neighborTile) ||
+                if (traversedTiles.Contains(neighborTile) ||
+                    // This is possible due to traversing to a tile that has null doorways
                     !LevelGenerationExtender.MeshContainers.ContainsKey(neighborTile))
                     continue;
 
-                DepthTesterQueue.Enqueue(new TileDepthTester(neighborTile, tile.iteration + 1));
+                depthTesterQueue.Enqueue(new TileDepthTester(neighborTile, tile.iteration + 1));
             }
         }
-    }
-
-    private static void OccludeByDistanceParallel()
-    {
-        Parallel.ForEach(LevelGenerationExtender.MeshContainers.Values, TestVisibility);
-    }
-
-    private static void OccludeByDistance()
-    {
-        foreach (var container in LevelGenerationExtender.MeshContainers.Values)
-            TestVisibility(container);
-    }
-
-    private static void TestVisibility(TileVisibility container)
-    {
-        if (VisibleTilesThisFrame.ContainsKey(container.parentTile)) return;
-
-        var position = container.parentTile.transform.position;
-
-        var shouldBeVisible = false;
-
-        if (!_depthTested)
-            shouldBeVisible = Vector3.SqrMagnitude(position - _playerPosition) <= SqrCullDistance;
-
-        foreach (var monitor in _enabledMonitors)
-        {
-            if (shouldBeVisible)
-                break;
-
-            shouldBeVisible |= Vector3.SqrMagnitude(position - monitor.targetedPlayer.transform.position) <= SqrCullDistance;
-        }
-
-        if (shouldBeVisible)
-            VisibleTilesThisFrame.TryAdd(container.parentTile, container);
     }
 }