+++ CullFactory/Plugin.cs
@@ -13,7 +13,7 @@ namespace CullFactory
 
         private const string Guid    = "com.fumiko.CullFactory";
         private const string Name    = "CullFactory";
-        private const string Version = "0.4.1";
+        private const string Version = "0.4.2";
 
         private void Awake()
         {
@@ -23,7 +23,7 @@ namespace CullFactory
             var harmony = new Harmony(Guid);
             harmony.PatchAll(typeof(LevelGenerationExtender));
 
-            Logger.LogInfo($"Plugin {Name} is loaded!");
+            Log($"Plugin {Name} is loaded!");
         }
 
         public static void Log(string s)